requirements.txt
Pillow
PyQt5
pytesseract
pywin32
PyYAML
win32gui
